nits: cleaning and doc
