+++ packages/sdk/src/messageBackend/index.ts
@@ -1,1 +1,3 @@
+
+export * from './goldsky';
 export * from './backend';
